Move function used in two files to its own file
